requirements.txt
apache-airflow==2.6.3
oss2==2.15.0
apache-airflow-providers-cncf-kubernetes==2.2.0
apache-airflow-providers-http==2.0.1
apache-airflow-providers-apache-hive==2.1.0
PyYAML==5.4.1
kubernetes==11.0.0
apache-airflow-providers-mysql==2.1.1
